Added Bullets, Arena Collision, Different Enemies, rounds, and Object Pooling

+++ Game1/Framework/Source/Random.cpp
@@ -15,7 +15,7 @@ int Random::RandInt(int min, int max)
 	return distribution(m_MiddleVariable);
 }
 
-float Random::RandFloat(int min, int max)
+float Random::RandFloat(float min, float max)
 {
 	std::uniform_real_distribution<float> distribution(min, max);
 	return distribution(m_MiddleVariable);
