+++ Game1/Framework/Source/Random.h
@@ -7,7 +7,7 @@ public:
 	Random();
 	~Random();
 	int RandInt(int min, int max);
-	float RandFloat(int min, int max);
+	float RandFloat(float min, float max);
 
 protected:
 	std::mt19937 m_MiddleVariable;
